UW_Python230_Internet_Programming_In_Python/PY230-django/blogging/serializers.py
from blogging.models import Post, Category
from rest_framework import serializers

class PostSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Post
        fields = ['title', 'text', 'author', 'created_date', 'modified_date', 'published_date', 'categories']


class CategorySerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Category
        fields = ['name', 'description']