UW_Python230_Internet_Programming_In_Python/PY230-django/requirements.txt
click==6.7
colorama==0.4.3
dj-database-url==0.5.0
Django==2.2.10
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
postgres==3.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
psycopg2-pool==1.1
pyconfiglib==1.3.3
Pygments==2.6.1
pypandoc==1.4
pytz==2019.3
superprompt==1.2.0
whitenoise==5.0.1
